add proajax script and modified firstimepurchase
