+++ app/Http/Kernel.php
@@ -19,6 +19,8 @@ class Kernel extends HttpKernel
         \App\Http\Middleware\TrimStrings::class,
         \Illuminate\Foundation\Http\Middleware\ConvertEmptyStringsToNull::class,
         \App\Http\Middleware\TrustProxies::class,
+        // This is our ajax middleware file that we are adding
+        \App\Http\Middleware\ProAjaxMiddleware::class,
     ];
 
     /**
